+++ seq_add.c
@@ -2,39 +2,38 @@
 #include <math.h>
 #include <stdlib.h>
 #include <stdbool.h>
-#include<unistd.h>
 #include <math.h>
 #include<string.h>
 #include<fcntl.h>
-void print_results(char* name,bool * primes_bool,unsigned long int min,unsigned long int max){
+void print_results(char* name,bool * primes_bool,unsigned long long int min,unsigned long long int max){
     FILE* x=fopen(name,"w+");
     fclose(x);   
-    for (unsigned long int i = min;i<=max;i++){
+    for (unsigned long long int i = min;i<=max;i++){
     FILE* f=fopen(name,"a");    
        
         
         if(!primes_bool[i]){
-            fprintf(f,"%ld\n",i);
+            fprintf(f,"%llu\n",i);
         }
         fclose(f);
     }}
  int main( int argc, char *argv[])
 {
-    unsigned long int min = 2;
-    unsigned long int max = 100000000;
-    unsigned long int size = max - min;
+    unsigned long long int min = 2;
+    unsigned long long int max = 100000000;
+    unsigned long long int size = max - min;
     FILE f;
     bool prime = false;
-    unsigned long int *primes;
+    unsigned long long int *primes;
     primes = malloc(sizeof(int) * size);
-    unsigned long int num_primes = 0;
+    unsigned long long int num_primes = 0;
     bool * primes_bool;
     primes_bool=malloc(sizeof(bool)*(max + 1));
     //memset(primes_bool, true, sizeof(primes_bool))false;
 
-    unsigned long int sqrt_sieve = sqrt(max);
-    unsigned long int tmp; // temporary sum
-    for (unsigned long int i = 2; i <= sqrt_sieve; i++)
+    unsigned long long int sqrt_sieve = sqrt(max);
+    unsigned long long int tmp; // temporary sum
+    for (unsigned long long int i = 2; i <= sqrt_sieve; i++)
     {
         if (primes_bool[i] == false)
         {
@@ -47,6 +46,6 @@ void print_results(char* name,bool * primes_bool,unsigned long int min,unsigned 
     }
 
     print_results("name",primes_bool,min,max);
-    printf("%ld",size);
+    printf("%llu",size);
 
 }
